+++ src/components/dialogSnackbar/dialogSnackbar.module.css
@@ -1,0 +1,4 @@
+.snackbar{
+  background-color: #f3e4a8 !important;
+  color: black !important;
+}
