+++ src/components/header/login/LoginModal.tsx
@@ -4,32 +4,14 @@ import styles from './loginModal.module.css';
 import CloseIcon from '@mui/icons-material/Close';
 import { useTranslation } from "react-i18next";
 import { PilasBloquesApi } from "../../../pbApi";
+import { DialogSnackbar } from "../../dialogSnackbar/DialogSnackbar";
 
-
-export interface DialogSnackbarProps {
-    open: boolean,
+export type LoginModalProps = {
+    open: boolean
     onClose: () => void
 }
 
-export const ServerErrorSnackbar:FC<DialogSnackbarProps> = ({open, onClose}) => {
-    const { t } = useTranslation('login');
-
-    return <Snackbar
-        anchorOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-        }}
-        open={open}
-        autoHideDuration={3000}
-        onClose={onClose}>
-            <SnackbarContent 
-                message={t('serverError')}
-                className={styles.serverError}
-            />
-        </Snackbar>
-}
-
-export const LoginModal:FC<DialogSnackbarProps> = ({open, onClose}) => {
+export const LoginModal:FC<LoginModalProps> = ({open, onClose}) => {
 
     const { t } = useTranslation('login');
 
@@ -89,9 +71,10 @@ export const LoginModal:FC<DialogSnackbarProps> = ({open, onClose}) => {
                         onChange={props => setPassword(props.target.value)}
                         required />
                     <Paper hidden={!wrongLogin} className={styles['paper']} elevation={2}>{t('wrong')}</Paper>
-                    <ServerErrorSnackbar 
+                    <DialogSnackbar 
                         open={serverError}
-                        onClose={() => setServerError(false)} />
+                        onClose={() => setServerError(false)} 
+                        message={t('serverError')}/>
                     <Button type='submit' className={styles['login-btn']}>{t('login')}</Button>
                     </form>
                     <a className={styles['link']} onClick={handleOnClose} href="#/password-recovery" target="">{t('forgot')}</a>
